fix: move components to respective folder
