+++ app/(site)/page.tsx
@@ -22,6 +22,13 @@ export default function Home() {
           </div>
         </div>
       </Header>
+
+      <div className="mt-2 mb-7 px-6">
+        <div className="flex justify-between items-center">
+          <h1 className="text-white text-2xl font-semibold">Newest Songs</h1>
+        </div>
+        <div>List of songs</div>
+      </div>
     </div>
   );
 }
